backend/app/requirements.txt
Django
djangorestframework
numpy 
pandas
scikit-learn
